refactor endpoint rule node step outputs
